feat: SessionRadio Modal 내 헤더
