+++ zoo-client/src/app/session-schedule/@modal/[id]/page.tsx
@@ -1,27 +1,14 @@
 'use client';
 
-import { PurchaseButton, RadioContent } from '@/components';
+import { SessionRadioModal } from '@/components';
 import SessionDetailModal from '@/components/common/modal/SessionDetailModal';
 import { useState } from 'react';
 
-import { useParams, useRouter } from 'next/navigation';
-import { SESSION_SCHEDULE_MESSAGES } from '@/constants/messages';
-import { useRadioStore } from '@/store/common/useRadioStore';
-
 export default function SessionModal() {
   const [showSecondModal, setShowSecondModal] = useState(false);
-  const id = useParams().id;
-  const { activeState } = useRadioStore();
-  const router = useRouter();
 
   const openSecondModal = () => {
     setShowSecondModal(true);
-  };
-
-  const sessionApplyHandler = () => {
-    console.log(activeState, id);
-    alert('세션 신청이 완료되었습니다!');
-    router.back();
   };
 
   return (
@@ -30,14 +17,7 @@ export default function SessionModal() {
         {!showSecondModal ? (
           <SessionDetailModal openSecondModal={openSecondModal} />
         ) : (
-          <div className="flex flex-col gap-10">
-            <RadioContent />
-            <PurchaseButton
-              func={sessionApplyHandler}
-              size={48}
-              text={SESSION_SCHEDULE_MESSAGES.buttonModalText}
-            />
-          </div>
+          <SessionRadioModal />
         )}
       </div>
     </div>
